CART-28 - Mejorando respuestas de casos particulares

+++ Spring/comercio/src/main/java/com/informatorio/comercio/controller/UsuarioController.java
@@ -1,16 +1,20 @@
 package com.informatorio.comercio.controller;
 import com.informatorio.comercio.domain.Usuario;
+import com.informatorio.comercio.dto.UsuarioCredencialesDto;
 import com.informatorio.comercio.repository.DireccionRepository;
 import com.informatorio.comercio.repository.UsuarioRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.format.annotation.DateTimeFormat;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
+import javax.validation.Valid;
 import java.util.Date;
 import java.util.List;
 
-import static com.informatorio.comercio.service.UsuarioService.modificarCredenciales;
 import static com.informatorio.comercio.service.UsuarioService.modificarDatos;
+import static org.springframework.http.HttpStatus.NOT_FOUND;
 
 @RestController()
 public class UsuarioController {
@@ -30,26 +34,65 @@ public class UsuarioController {
     }
 
     @GetMapping(value = "/usuario/{id}")
-    public Usuario verUsuario(@PathVariable("id") Long id){
-        return usuarioRepository.getById(id);
+    public Object verUsuario(@PathVariable("id") Long id){
+        Usuario usuario =  usuarioRepository.findById(id).orElse(null);
+        if (usuario == null){
+            return new ResponseEntity<>("El usuario con el id indicado no existe.",HttpStatus.CONFLICT);
+        }else{
+            return new ResponseEntity<>(usuario,HttpStatus.OK);
+        }
     }
 
     @DeleteMapping(value = "/usuario/{id}")
-    public void borrarUsuario(@PathVariable("id") Long id){
-        Usuario usuario =  usuarioRepository.getById(id);
-        usuarioRepository.delete(usuario);
+    public Object borrarUsuario(@PathVariable("id") Long id){
+        Usuario usuario =  usuarioRepository.findById(id).orElse(null);
+        if (usuario == null){
+            return new ResponseEntity<>("El usuario con el id indicado no existe.",HttpStatus.CONFLICT);
+        }else{
+            if (usuario.getOrdenes().size()>=1){
+                usuario.setCambiarEstado();
+                return new ResponseEntity<>("El usuario tiene compras registradas, pero ahora su cuenta esta inactiva.",HttpStatus.OK);
+            }else{
+                usuarioRepository.delete(usuario);
+                return new ResponseEntity<>("El usuario fue eliminado.",HttpStatus.OK);
+            }
+        }
     }
 
     @PutMapping(value = "/usuario/{id}")
-    public Usuario modificarUsuario(@PathVariable("id") Long id, @RequestBody Usuario usuario){
-        Usuario user = usuarioRepository.getById(id);
-        return modificarDatos(user, usuario);
+    public ResponseEntity<?> modificarUsuario(@PathVariable("id") Long id, @Valid @RequestBody Usuario usuario){
+        Usuario user = usuarioRepository.findById(id).orElse(null);
+        if (user == null){
+            return new ResponseEntity<>("No existe usuario con el id ingresado.",NOT_FOUND);
+        }else{
+            return modificarDatos(user, usuario);
+        }
     }
 
-    @PutMapping(value="/usuario/{id}/credencial")
-    public Usuario modificarUsuarioCredenciales(@PathVariable("id") Long id, @RequestBody Usuario usuario){
-        Usuario user = usuarioRepository.getById(id);
-        return modificarCredenciales(user, usuario);
+    @PutMapping(value="/usuario/{id}/credencial/contrasenia")
+    public Object modificarUsuarioCredencialesPwd(@PathVariable("id") Long id,@RequestBody String contrasenia){
+        Usuario user = usuarioRepository.findById(id).orElse(null);
+        if (user == null){
+            return new ResponseEntity<>("No existe usuario con el id ingresado.",NOT_FOUND);
+        }else{
+            user.setPassword(contrasenia);
+            usuarioRepository.save(user);
+            user = usuarioRepository.findById(id).orElse(null);
+            return new ResponseEntity<>(user,HttpStatus.CREATED);
+        }
+    }
+
+    @PutMapping(value="/usuario/{id}/credencial/email")
+    public Object modificarUsuarioCredencialesEmail(@PathVariable("id") Long id,@RequestBody String email){
+        Usuario user = usuarioRepository.findById(id).orElse(null);
+        if (user == null){
+            return new ResponseEntity<>("No existe usuario con el id ingresado.",NOT_FOUND);
+        }else{
+            user.setEmail(email);
+            usuarioRepository.save(user);
+            user = usuarioRepository.findById(id).orElse(null);
+            return new ResponseEntity<>(user,HttpStatus.CREATED);
+        }
     }
 
     @GetMapping(value = "/usuario/direccion")
